feat: enhance region handling in Header and add region management actions
