+++ src/routes/projects/[project]/+page.server.ts
@@ -1,0 +1,34 @@
+import { regionUrl } from '$lib/url';
+import { addRegionSchema } from '$routes/projects/[project]/regions/[region]/schema';
+import { redirect, type Actions } from '@sveltejs/kit';
+import { fail, setError, superValidate } from 'sveltekit-superforms';
+import { zod } from 'sveltekit-superforms/adapters';
+
+export const actions: Actions = {
+	addRegion: async ({ request, params, locals }) => {
+		const { project } = params;
+		const addRegionForm = await superValidate(request, zod(addRegionSchema));
+
+		const projectData = locals.userState.projects.find((p) => p.name === project);
+		if (!projectData) {
+			return fail(404, { addRegionForm, error: `Project "${project}" not found` });
+		}
+
+		const isRegionDuplicate = projectData.regions.some((r) => r.name === addRegionForm.data.name);
+		if (isRegionDuplicate) {
+			setError(addRegionForm, 'name', 'Region names must be unique');
+		}
+		if (!addRegionForm.valid) {
+			return fail(400, { addRegionForm });
+		}
+
+		projectData.regions.push({
+			name: addRegionForm.data.name,
+			formValues: {},
+			hasRunBaseline: false,
+			cases: []
+		});
+
+		return redirect(303, regionUrl(projectData.name, addRegionForm.data.name));
+	}
+};
